Module

+++ System.Module/ClassIndex.cs
@@ -6,5 +6,5 @@ namespace System.Module;
 
 public class ClassIndex : InfraObject
 {
-    public ulong Value { get; set; }
+    public int Value { get; set; }
 }
